+++ code/src/sixsq/nuvla/server/app/server.clj
@@ -15,9 +15,9 @@
     [sixsq.nuvla.server.middleware.authn-info :refer [wrap-authn-info]]
     [sixsq.nuvla.server.middleware.base-uri :refer [wrap-base-uri]]
     [sixsq.nuvla.server.middleware.cimi-params :refer [wrap-cimi-params]]
-    [sixsq.nuvla.server.middleware.gzip :refer [wrap-gzip-uncompress]]
     [sixsq.nuvla.server.middleware.default-content-type :refer [default-content-type]]
     [sixsq.nuvla.server.middleware.exception-handler :refer [wrap-exceptions]]
+    [sixsq.nuvla.server.middleware.gzip :refer [wrap-gzip-uncompress]]
     [sixsq.nuvla.server.middleware.logger :refer [wrap-logger]]
     [sixsq.nuvla.server.middleware.redirect-cep :refer [redirect-cep]]
     [sixsq.nuvla.server.resources.common.dynamic-load :as resources]
@@ -40,7 +40,6 @@
   (-> (routes/get-main-routes)
 
       ;; handler/site
-      wrap-gzip-uncompress
       wrap-cimi-params
       wrap-base-uri
       wrap-keyword-params
@@ -49,6 +48,7 @@
       wrap-exceptions
       wrap-authn-info
       (wrap-json-body {:keywords? true})
+      wrap-gzip-uncompress
       (wrap-json-response {:pretty           true
                            :escape-non-ascii true})
       (default-content-type "application/json")