+++ code/src/sixsq/nuvla/server/middleware/logger.clj
@@ -44,22 +44,6 @@
     (display-elapsed-time-millis start current-time-millis)
     formatted-request))
 
-(defn format-request-debug
-  [request]
-  (display-space-separated
-    (-> request :request-method name str/upper-case)
-    (:uri request)
-    request))
-
-
-(defn format-response-debug
-  [formatted-request response start current-time-millis]
-  (display-space-separated
-    (:status response)
-    (display-elapsed-time-millis start current-time-millis)
-    formatted-request
-    response))
-
 
 (defn log-response
   [status formatted-message]
@@ -78,8 +62,8 @@
   [handler]
   (fn [request]
     (let [start             (System/currentTimeMillis)
-          formatted-request (format-request-debug request)
+          formatted-request (format-request request)
           _                 (log/debug formatted-request)
           {:keys [status] :as response} (handler request)
-          _                 (log-response status (format-response-debug formatted-request response start (System/currentTimeMillis)))]
+          _                 (log-response status (format-response formatted-request response start (System/currentTimeMillis)))]
       response)))